Issue 7696 - Move add-table request and response types into core
