+++ java/core/src/test/java/sleeper/core/table/AddTableResponseSerDeTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.restapi.addTable;
+package sleeper.core.table;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
