+++ java/core/src/test/java/sleeper/core/table/example.approved.json
