+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/instance/AwsSleeperTablesDriver.java
@@ -36,11 +36,11 @@ import sleeper.core.properties.table.TableProperties;
 import sleeper.core.properties.table.TablePropertiesProvider;
 import sleeper.core.properties.table.TablePropertiesStore;
 import sleeper.core.statestore.StateStoreProvider;
+import sleeper.core.table.AddTableRequest;
+import sleeper.core.table.AddTableRequestSerDe;
+import sleeper.core.table.AddTableResponse;
+import sleeper.core.table.AddTableResponseSerDe;
 import sleeper.core.table.TableIndex;
-import sleeper.restapi.addTable.AddTableRequest;
-import sleeper.restapi.addTable.AddTableRequestSerDe;
-import sleeper.restapi.addTable.AddTableResponse;
-import sleeper.restapi.addTable.AddTableResponseSerDe;
 import sleeper.statestore.StateStoreFactory;
 import sleeper.systemtest.drivers.util.SystemTestClients;
 import sleeper.systemtest.dsl.instance.SleeperTablesDriver;
